Refactor a bit and move backend selection to features folder
